first steps towards usbkbd/hdmi for g12a and sm1
at least on my g12a box its still a bit unreliable if i get the
prompt etc. on hdmi or not, on my sm1 box it looks more reliable.
maybe its just related to those boxes. also it is still open if
the serial console can remain enabled of if it has to get disabled
like on gxl and gxbb due to noise input from the serial line.
right now its still enabled for the usbkdb/hdmi case and prints
output to serial, but input is only taken from usbkbd somehow.